+++ python/metatensor_core/tests/labels.py
@@ -1,3 +1,5 @@
+import re
+
 import numpy as np
 import pytest
 
@@ -430,3 +432,24 @@ def test_equal_non_contiguous():
     labels_non_contiguous = Labels(names=["a", "b"], values=values_non_contiguous)
     labels_contiguous = Labels(names=["a", "b"], values=values_contiguous)
     assert labels_contiguous == labels_non_contiguous
+
+
+def test_ownership_transfer():
+    """Test releasing and recovering Labels via ``unsafe_from_ptr``"""
+    labels = Labels(names=["foo", "bar"], values=np.array([[1, 2], [3, 4], [5, 6]]))
+
+    raw = labels.release()
+    message = "can not access these Labels, they have been released"
+    with pytest.raises(RuntimeError, match=re.escape(message)):
+        labels.as_mts_labels_t()
+
+    recovered = Labels.unsafe_from_ptr(raw)
+    assert recovered == Labels(
+        names=["foo", "bar"], values=np.array([[1, 2], [3, 4], [5, 6]])
+    )
+
+    raw = recovered.release()
+    with pytest.raises(RuntimeError, match=re.escape(message)):
+        recovered.as_mts_labels_t()
+
+    Labels.unsafe_from_ptr(raw)