+++ metatensor-core/CHANGELOG.md
@@ -17,7 +17,9 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 #### Removed
 -->
 
-### Added
+### metatensor-core C++
+
+#### Added
 
 - `TensorMap::unsafe_from_ptr` to replace the `TensorMap(mts_tensormap_t*)`
   constructor, for symmetry with the `TensorBlock` API. The constructor is now
@@ -28,6 +30,17 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
   TensorMap, as well as `TensorMap::is_view` and `TensorBlock::is_view`.
 - `Labels::release`, `TensorBlock::release`, and `TensorMap::release` to take
   back ownership of the underlying C API pointers.
+
+### metatensor-core Python
+
+#### Added
+
+- `TensorMap.unsafe_from_ptr`, `TensorBlock.unsafe_from_ptr` and
+  `Labels.unsafe_from_ptr` to create Python classes from raw ctypes pointers
+- `TensorMap.unsafe_view_from_ptr` and `TensorBlock.unsafe_view_from_ptr` to
+  create non-owning views of existing raw pointers.
+- `Labels.release`, `TensorBlock.release`, and `TensorMap.release` to take
+  back ownership of the underlying pointers.
 
 ## [Version 0.2.0](https://github.com/metatensor/metatensor/releases/tag/metatensor-core-v0.2.0) - 2026-05-13